rename functional to integration
